refactor: update project structure and dependencies; remove unused client files
