--- client/src/components/NotificationsMenu.jsx
@@ -1,188 +1,0 @@
-import { useState, useEffect } from 'react';
-import {
-  IconButton,
-  Badge,
-  Menu,
-  MenuItem,
-  Typography,
-  Box,
-  Button,
-  CircularProgress,
-  List,
-  ListItem,
-  ListItemText,
-  Divider
-} from '@mui/material';
-import {
-  Notifications as NotificationsIcon,
-  Circle as CircleIcon
-} from '@mui/icons-material';
-import { format } from 'date-fns';
-
-export default function NotificationsMenu({ userId = '123' }) {
-  const [anchorEl, setAnchorEl] = useState(null);
-  const [notifications, setNotifications] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState('');
-
-  const fetchNotifications = async () => {
-    if (!userId) return;
-    
-    setLoading(true);
-    setError('');
-    try {
-      const response = await fetch(`/api/notifications?userId=${userId}`);
-      if (!response.ok) {
-        throw new Error('Failed to fetch notifications');
-      }
-      const data = await response.json();
-      setNotifications(Array.isArray(data) ? data : []);
-    } catch (err) {
-      console.error('Error fetching notifications:', err);
-      setError('Failed to load notifications');
-      setNotifications([]);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchNotifications();
-    // Set up polling for notifications
-    const interval = setInterval(fetchNotifications, 30000); // Poll every 30 seconds
-    return () => clearInterval(interval);
-  }, [userId]);
-
-  const handleClick = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const handleMarkAllRead = async () => {
-    if (!userId || notifications.length === 0) return;
-
-    try {
-      const promises = notifications.map(notification =>
-        fetch('/api/notifications', {
-          method: 'PUT',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            notificationId: notification.id,
-            userId
-          }),
-        })
-      );
-
-      await Promise.all(promises);
-      fetchNotifications();
-    } catch (err) {
-      console.error('Error marking notifications as read:', err);
-      setError('Failed to mark notifications as read');
-    }
-  };
-
-  const getNotificationIcon = (type) => {
-    switch (type) {
-      case 'MEETING_INVITATION':
-        return '🤝';
-      case 'MEETING_UPDATED':
-        return '📝';
-      case 'MEETING_CANCELLED':
-        return '❌';
-      case 'MEETING_REMINDER':
-        return '⏰';
-      default:
-        return '📢';
-    }
-  };
-
-  const notificationCount = notifications?.length || 0;
-
-  return (
-    <>
-      <IconButton
-        color="inherit"
-        onClick={handleClick}
-        aria-label={`${notificationCount} new notifications`}
-      >
-        <Badge badgeContent={notificationCount} color="error">
-          <NotificationsIcon />
-        </Badge>
-      </IconButton>
-
-      <Menu
-        anchorEl={anchorEl}
-        open={Boolean(anchorEl)}
-        onClose={handleClose}
-        PaperProps={{
-          sx: {
-            width: 360,
-            maxHeight: 400,
-          },
-        }}
-      >
-        <Box sx={{ p: 2, display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-          <Typography variant="h6">Notifications</Typography>
-          {notificationCount > 0 && (
-            <Button size="small" onClick={handleMarkAllRead}>
-              Mark all as read
-            </Button>
-          )}
-        </Box>
-
-        <Divider />
-
-        {loading ? (
-          <Box sx={{ display: 'flex', justifyContent: 'center', p: 2 }}>
-            <CircularProgress size={24} />
-          </Box>
-        ) : error ? (
-          <Box sx={{ p: 2 }}>
-            <Typography color="error">{error}</Typography>
-          </Box>
-        ) : notificationCount === 0 ? (
-          <Box sx={{ p: 2 }}>
-            <Typography color="text.secondary" align="center">
-              No new notifications
-            </Typography>
-          </Box>
-        ) : (
-          <List sx={{ p: 0 }}>
-            {notifications.map((notification) => (
-              <ListItem
-                key={notification.id}
-                sx={{
-                  borderBottom: '1px solid',
-                  borderColor: 'divider',
-                  '&:hover': {
-                    backgroundColor: 'action.hover',
-                  },
-                }}
-              >
-                <Box sx={{ mr: 2 }}>{getNotificationIcon(notification.type)}</Box>
-                <ListItemText
-                  primary={notification.title}
-                  secondary={
-                    <>
-                      <Typography variant="body2" component="span" display="block">
-                        {notification.message}
-                      </Typography>
-                      <Typography variant="caption" color="text.secondary">
-                        {format(new Date(notification.createdAt), 'PPp')}
-                      </Typography>
-                    </>
-                  }
-                />
-              </ListItem>
-            ))}
-          </List>
-        )}
-      </Menu>
-    </>
-  );
-}
