--- client/src/components/UpdateMeetingForm.jsx
@@ -1,154 +1,0 @@
-import { useState, useEffect } from 'react';
-import {
-  Box,
-  TextField,
-  Button,
-  Stack,
-  Alert,
-  FormControl,
-  InputLabel,
-  Select,
-  MenuItem,
-  CircularProgress,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions
-} from '@mui/material';
-import { DateTimePicker } from '@mui/x-date-pickers';
-import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
-import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
-
-const durations = [
-  { value: 15, label: '15 minutes' },
-  { value: 30, label: '30 minutes' },
-  { value: 45, label: '45 minutes' },
-  { value: 60, label: '1 hour' },
-  { value: 90, label: '1.5 hours' },
-  { value: 120, label: '2 hours' },
-];
-
-export default function UpdateMeetingForm({ meeting, open, onClose, onUpdate }) {
-  const [formData, setFormData] = useState({
-    title: '',
-    description: '',
-    startTime: null,
-    duration: 30,
-    status: 'scheduled'
-  });
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState('');
-
-  useEffect(() => {
-    if (meeting) {
-      setFormData({
-        title: meeting.title,
-        description: meeting.description || '',
-        startTime: new Date(meeting.startTime),
-        duration: meeting.duration,
-        status: meeting.status
-      });
-    }
-  }, [meeting]);
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setError('');
-    setLoading(true);
-
-    try {
-      if (!formData.title || !formData.startTime) {
-        throw new Error('Please fill in all required fields');
-      }
-
-      await onUpdate(formData);
-      onClose();
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <Dialog open={open} onClose={onClose} maxWidth="sm" fullWidth>
-      <DialogTitle>Update Meeting</DialogTitle>
-      <DialogContent>
-        <Box component="form" onSubmit={handleSubmit} sx={{ mt: 2 }}>
-          <Stack spacing={3}>
-            {error && <Alert severity="error">{error}</Alert>}
-
-            <TextField
-              label="Meeting Title"
-              required
-              value={formData.title}
-              onChange={(e) => setFormData(prev => ({ ...prev, title: e.target.value }))}
-              disabled={loading}
-            />
-
-            <TextField
-              label="Description"
-              multiline
-              rows={3}
-              value={formData.description}
-              onChange={(e) => setFormData(prev => ({ ...prev, description: e.target.value }))}
-              disabled={loading}
-            />
-
-            <LocalizationProvider dateAdapter={AdapterDateFns}>
-              <DateTimePicker
-                label="Start Time"
-                value={formData.startTime}
-                onChange={(newValue) => setFormData(prev => ({ ...prev, startTime: newValue }))}
-                disabled={loading}
-                disablePast
-                required
-              />
-            </LocalizationProvider>
-
-            <FormControl>
-              <InputLabel>Duration</InputLabel>
-              <Select
-                value={formData.duration}
-                onChange={(e) => setFormData(prev => ({ ...prev, duration: e.target.value }))}
-                disabled={loading}
-                label="Duration"
-              >
-                {durations.map(({ value, label }) => (
-                  <MenuItem key={value} value={value}>
-                    {label}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-
-            <FormControl>
-              <InputLabel>Status</InputLabel>
-              <Select
-                value={formData.status}
-                onChange={(e) => setFormData(prev => ({ ...prev, status: e.target.value }))}
-                disabled={loading}
-                label="Status"
-              >
-                <MenuItem value="scheduled">Scheduled</MenuItem>
-                <MenuItem value="cancelled">Cancelled</MenuItem>
-                <MenuItem value="completed">Completed</MenuItem>
-              </Select>
-            </FormControl>
-          </Stack>
-        </Box>
-      </DialogContent>
-      <DialogActions>
-        <Button onClick={onClose}>Cancel</Button>
-        <Button
-          onClick={handleSubmit}
-          variant="contained"
-          disabled={loading}
-          startIcon={loading ? <CircularProgress size={20} /> : null}
-        >
-          Update Meeting
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
-}
