+++ next.config.mjs
@@ -1,4 +1,6 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  // App directory is stable in Next.js 15
+};
 
 export default nextConfig;
