+++ package.json
@@ -22,19 +22,18 @@
     "date-fns": "^2.30.0",
     "dotenv": "^16.4.7",
     "mysql2": "^3.7.1",
-    "next": "14.0.4",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "^15.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "sequelize": "^6.37.5",
     "sqlite3": "^5.1.7",
     "uuid": "^11.0.4"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "eslint": "^9",
-    "eslint-config-next": "15.1.3",
+    "autoprefixer": "^10.0.1",
     "postcss": "^8",
-    "sequelize-cli": "^6.6.2",
-    "tailwindcss": "^3.4.1"
+    "tailwindcss": "^3.3.0",
+    "eslint": "^8",
+    "eslint-config-next": "^15.0.0"
   }
 }
